+++ PHEngine/PHEngine/src/Core/GameCore/TextHandler.h
@@ -28,7 +28,7 @@ public:
 
     std::shared_ptr<HudTextField> CreateTextField(
         const std::string& fontName,
-        const float fontSize,
+        const int32_t fontSize,
         const std::string& text,
         const glm::vec3& color,
         const glm::vec2& position,
@@ -39,7 +39,7 @@ public:
 
     std::shared_ptr<HudTextField> CreateEmptyTextField(
         const std::string& fontName,
-        const float fontSize,
+        const int32_t fontSize,
         const glm::vec3& color,
         const bool receiveUpdateOnTextScreenSpaceSizeChanged,
         const float lineMaxSize,
